# tags: improve performance for small tag sets