TLDR: I had some time to kill on flight last week and noticed that many tag operations occur with 1 or 2 tags - so I made that roughly ~40% faster and reduced allocs and memory usage by roughly 2/3, which is what I really care about.

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkSerializeTags-8         444           445           +0.23%
BenchmarkSerializeTags_One-8     162           94.0          -41.98%
BenchmarkSerializeTags_Two-8     237           145           -38.82%

benchmark                        old allocs     new allocs     delta
BenchmarkSerializeTags-8         3              3              +0.00%
BenchmarkSerializeTags_One-8     3              1              -66.67%
BenchmarkSerializeTags_Two-8     3              1              -66.67%

benchmark                        old bytes     new bytes     delta
BenchmarkSerializeTags-8         272           272           +0.00%
BenchmarkSerializeTags_One-8     96            32            -66.67%
BenchmarkSerializeTags_Two-8     128           32            -75.00%
```